Auto-assign Snake Trap pets to owner
The summon is done by an invisible trigger that is not tied to the caster. We assign the snakes by selecting the first 8-14 summones (depending on the spec, as Survival is the only one that should play 3/3 Trap Mastery) that happens within the first 30 seconds from the SPELL_CREATE event that is fired when the trap is set. This could lead to issues when multiple Snake Traps from different Hunters are active at the same time (e.g. when two are cast but only one is triggered, we cannot know which one it is), but we don't think this can be resolved