Parse actors by encounter; improve pet assignment
Actors are now assigned the encounter(s) they participated in through the corresponding starting time; this makes fetching valid units for a given encounter faster and easier.
The pet assignment tool now allows selection by encounter so it is easier to assign the correct pets without having to rely on impractical things such as GUID.
The new pet owner selection menu is also updated to show a class icon, in particular when selecting an encounter the correct spec icon is selected.
Closing the pet assignment window automatically refreshes the table.